+++ src/main/java/roomescape/exception/MissingParameterException.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.exception;
 
 public class MissingParameterException extends RuntimeException {
 }
